blog_project/requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.3.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
